+++ R/resmush_file.R
@@ -37,7 +37,7 @@
 #' @family optimize
 #' @export
 #' @encoding UTF-8
-#' @examplesIf httr2::is_online()
+#' @examplesIf resmush_is_online()
 #'
 #' \donttest{
 #' png_file <- system.file("extimg/example.png", package = "resmush")
@@ -115,9 +115,7 @@ resmush_file_single <- function(
   outfile <- add_suffix(file, suffix, overwrite)
   res <- new_resmush_result(file)
 
-  # Allow checks to simulate offline mode.
-  test <- getOption("resmush_test_offline", FALSE)
-  if (any(isFALSE(httr2::is_online()), test)) {
+  if (isFALSE(resmush_is_online())) {
     res$notes <- "Offline"
     return(invisible(res))
   }
@@ -137,12 +135,10 @@ resmush_file_single <- function(
     return(invisible(res))
   }
 
-  # nocov start
   if (!"dest" %in% names(res_post)) {
     res$notes <- "API not responding, check https://resmush.it/status"
     return(invisible(res))
   }
-  # nocov end
 
   dwn_opt <- download_optimized_file(
     url = res_post$dest,
@@ -154,9 +150,7 @@ resmush_file_single <- function(
     return(NULL)
   }
 
-  # Allow checks to simulate an invalid download response.
-  test_corner <- getOption("resmush_test_corner", FALSE)
-  if (any(httr2::resp_status(dwn_opt) != 200, test_corner)) {
+  if (httr2::resp_status(dwn_opt) != 200) {
     res$notes <- "API not responding, check https://resmush.it/status"
     return(invisible(res))
   }